Change to systemd-activated dbus service
As a system-wide dbus service, it's not properly running with user 'app'
privilege if the login user isn't it. For example, in multi-user
environment.

Since it works perfectly with non-root privilege, so it's not reasonable
to change "User=" to root.

Given that meanwhile it's a systemd user session service, so it will
running with the login user if it started by systemd, so it works in
multi-user environment. So this patch looks straight forward.

Change-Id: Ic970bc5c00a162f49e716a231d8d0c06a15d2412
Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>