Add the device lists of scanning to fix the crash issue
Add the device lists of scanning.
So that providing the device lists when UI is needed.

Bluetooth-frwk is stable when it own the list by itself.

If getting from Bluez. Bluetooth-frwk will crash sometime.
The root cause is  message format issues.

Thefefore, Bluetooth-frwk owner the device lists.
Bluetooth-frwk will be more stable.

Change-Id: I812d616947bbe00e594bb3b691ea830e81e22990