Obex/Notify: Create transfer notify for global watched notifier
transfer notify should be as a packet, when you send notify by g_idle_add_full.
Global notifier hold data include transfer pointer, if the notify data is
modified before g_main_loop send that notify. The packet that is notify data
may be not you want. We should create a new notify for global notifier, keeping
the data as a packet to send.

Change-Id: Iaf954ba619d3e4d3309eae14f62145208bae3ab9
Signed-off-by: Wu Jiangbo <jiangbox.wu@intel.com>