add Tizen Common plugin
In this version, popups content is filed in notification db.
Then a notification client has to display the popup and reply to
bluetooth-service daemon using draft bt_agent_xxx sync API.

In addition, the pairing timeout was increased to 20 sec and
a bluetooth icon was added.

Compared to previous patchset :
- MOBILE and COMMON build flags were prefixed with 'TIZEN_'
- a draft bt_agent_xxx sync API is proposed.
	It replaces the "ReplyConfirmation" method.
- 3 popup types were defined and add to common plugin

Change-Id: Ib38f7baf62b0d4c128df1700263bc4d9b2c36af6
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>