adapt NTB to Tizen 3
This patch aims to send popup data to notification-service.
Then a notification-service client is supposed to subscribe to bt_agent
notifications in order to display them and handle user replies.

This approach is not compliant with proposed bt_agent CAPI which is based
on dbus callbacks. It is not possible to catch new user popups by dbus
callbacks and by notification-service at the same time.
To use notification-service we need a synchronous bt_agent API for user
replies.

Change-Id: Ib38f7baf62b0d4c128df1700263bc4d9b2c36af6
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>