add Tizen Common plugin
In this version, popups content is filed in notification db.
Then a notification client has to display the popup and reply to
bluetooth-service using 'ReplyConfirmation' dbus method call.
When bluetooth-service catches the 'ReplyConfirmation' method call,
we release 'RequestConfirmation' dbus call.

In addition, the pairing timeout was increased to 20 sec and
a bluetooth icon was added.

Change-Id: Ib38f7baf62b0d4c128df1700263bc4d9b2c36af6
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>