Fix the remove bt device crash issue
In the device removed process,
object->interfaces is in the circle.

However, the device destroy function remove the list device
in object->interfaces.

It is strength and will result in the program crash.
Therefore, fix it.

At the same time, fix the related adapter and gatt too.

And remove the unuseful source code.

Change-Id: I17e26118b1f17cc1d3e316f29e360b889978b7f0
Signed-off-by: Wu Zheng <wu.zheng@intel.com>