use bluez dbus api to enable/disable adapter
It was previously done invoking bluetooth-tools scripts.
For example, to disable adapter, one of the script kills bluez daemon process.
Now we just modify 'Powered' adapter property.

The patch also removes the timeout callback when enabling adapter since it can
disable adapter if the status is not correct. This adapter status is handled
by bluez4 adapter dbus api code which needs to be adapted.

Change-Id: I4bb5c4c03462a7fe9957e32dd8b196d37f358d30
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>