store bt user privileges in /usr/share/bluetooth-frwk directory
HOME environment variable is not forwarded in bluetooth-service process context when
this daemon is spawned by dbus activation. So bluetooth-service daemon was not able
to create .bt_userprivileges file in home directory.

As bluetooth-service daemon will be owned by a dedicated user, there is no home
directory, and so let's move it to /usr/share.

This patch replaces 'app' user by 'guest' user in order to run it on Tizen Common
which doesn't have 'app' user.
It also handles smack labels and directory ownership.

Bug-Tizen: TC-1981

Change-Id: I80968817ca5872b6b79b22ed120821e7987a1846
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>