keep current bluetooth state when starting bt-service
Previsouly when bt-service was started, it always tries to enable the bluetooth.
Therefore when the bluetooth is already enabled when starting bt-service,
it was not clean.

This patch allows to keep consistency between bluez and tizen bluetooth states.
By this way, if bluetooth is enabled when starting bt-service, bt-service just set vconf
and other variables. Following same logic, if bluetooth is down when starting
bt-service, we keep it down and align vconf status.

Change-Id: Iaf9170a739aa4a4c0ac37f1d98caafea1df50643
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>