fix destroy bonding device use case
9319c8 "add device disappear event" commit brings a regression.
It added a logic of storing paired device in glist which is not good and check
paired device in glist before sending a 'destroy bond' or 'disappear device' events.

It results that if a device was already paired before starting bt-service daemon,
the paired device was not listed in the glist and so 'destroy bond' event
was never triggered in case of unbonding request.

To fix it, let's consider that if device is present in glist, we trig 'disappear device'
event and otherwise we trig 'destroy bond' event.

Change-Id: I202cdeea48af7e3168c089c4164c7d7ef4711b2c
Signed-off-by: Corentin Lecouvey <corentin.lecouvey@open.eurogiciel.org>